Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -22,7 +22,7 @@ int main()
     case '/':
         if (num2 != 0)
         {
-            printf("%f\n", (float)num1 / num2);
+            printf("%f\n", (int)num1 / num2);
         }
         else
         {
